--- strelets/provision_virtual_chain.go
@@ -1,1 +1,0 @@
-package strelets
